+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/notification/RecruitmentNotificationSideEffect.kt
@@ -2,6 +2,6 @@ package `in`.koreatech.koin.feature.recruitment.ui.notification
 
 internal sealed class RecruitmentNotificationSideEffect {
     data object Error : RecruitmentNotificationSideEffect()
-    data class NavigateToApplicantManagement(val recruitmentId: Long) : RecruitmentNotificationSideEffect()
+    data class NavigateToApplicantManagement(val recruitmentId: Int) : RecruitmentNotificationSideEffect()
     data object Deleted : RecruitmentNotificationSideEffect()
 }
